requirements.txt
pyperclip
alpaca-trade-api
python-dotenv
yfinance
pandas
